server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
